+++ DigitaleBriefwahl/Mail/MapiEmailProvider.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Eberhard Beilharz
+// Copyright (c) 2021-2022 Eberhard Beilharz
 // This software is licensed under the GNU General Public License version 3
 // (https://opensource.org/licenses/GPL-3.0)