+++ DigitaleBriefwahl/Mail/OutlookEmailProvider.cs
@@ -1,5 +1,6 @@
-﻿// // Copyright (c) 2018 SIL International
-// // This software is licensed under the MIT License (http://opensource.org/licenses/MIT)
+﻿// Copyright (c) 2018-2022 Eberhard Beilharz
+// This software is licensed under the GNU General Public License version 3
+// (https://opensource.org/licenses/GPL-3.0)
 
 using System;
 using System.Collections.Generic;
@@ -21,19 +22,9 @@ namespace DigitaleBriefwahl.Mail
 
 		protected override string FormatString => "-c IPM.Note /m \"{0}{1}{2}&subject={3}&body={4}\"{5}";
 
-		protected override string EmailCommand
-		{
-			get
-			{
-				// we already checked that the mail client is Outlook
-				using var key = Registry.ClassesRoot.OpenSubKey(@"mailto\shell\open\command");
-				var value = key?.GetValue("") as string;
-				var regex = new Regex(@"^""([^""]+)""");
-				if (value == null || !regex.IsMatch(value))
-					return null;
-				return regex.Match(value).Groups[1].Value;
-			}
-		}
+		protected override string EmailCommand =>
+			// we already checked that the mail client is Outlook
+			MailtoCommand;
 
 		private static string GetArguments(IEnumerable<string> arguments)
 		{